Custom error page
When the backend is not healthy we jump into this else condition. The first thing Varnish will do is deliver us stale content, but when we pass the grace time, users will see an error page. I think it is better to configure a custom error page (with customized text) then the default error page, which is kinda ugly.